% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ch.distances}
\alias{ch.distances}
\title{Create distances data from matrix or file.}
\usage{
ch.distances(matrix = NA, file = NA)
}
\arguments{
\item{matrix}{Symmetric distance matrix. Row and column names are required and used as item ids.}

\item{file}{File from which to read the distance matrix.}
}
\description{
Specify either a symmetric distance matrix or the file from which to read the matrix.
}

