% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corehunter-package.R
\docType{package}
\name{corehunter-package}
\alias{corehunter}
\alias{corehunter-package}
\title{corehunter: Multi-Purpose Core Subset Selection}
\description{
Core Hunter is a tool to sample diverse, representative subsets from large germplasm collections, with minimum redundancy. Such so-called core collections have applications in plant breeding and genetic resource management in general. Core Hunter can construct cores based on genetic marker data, phenotypic traits or precomputed distance matrices, optimizing one of many provided evaluation measures depending on the precise purpose of the core (e.g. high diversity, representativeness, or allelic richness). In addition, multiple measures can be simultaneously optimized as part of a weighted index to bring the different perspectives closer together. The Core Hunter library is implemented in Java 8 as an open source project (see \url{http://www.corehunter.org}).
}
\examples{
\dontrun{
# sample core based on genetic marker data (default format)
geno.file <- system.file("extdata", "genotypes.csv", package = "corehunter")
geno <- genotypes(file = geno.file, format = "default")
core <- sampleCore(geno)

# sample core based on genetic marker data (biparental format)
geno.file <- system.file("extdata", "genotypes-biparental.csv", package = "corehunter")
geno <- genotypes(file = geno.file, format = "biparental")
core <- sampleCore(geno)

# sample core based on genetic marker data (frequency format)
geno.file <- system.file("extdata", "genotypes-frequency.csv", package = "corehunter")
geno <- genotypes(file = geno.file, format = "frequency")
core <- sampleCore(geno)

# sample core based on phenotypic traits
pheno.file <- system.file("extdata", "phenotypes.csv", package = "corehunter")
pheno <- phenotypes(file = pheno.file)
core <- sampleCore(pheno)

# sample core based on precomputed distance matrix
dist.file <- system.file("extdata", "distances.csv", package = "corehunter")
dist <- distances(file = dist.file)
core <- sampleCore(dist)

# sample core from genotypes with custom objective (allelic richness)
core <- sampleCore(geno, obj = objective("HE"))

# sample core from genotypes with custom size and objective (representativeness)
core <- sampleCore(geno, obj = objective("AN", "MR"), size = 0.1)

# sample core from genotypes with custom size and stop condition
core <- sampleCore(geno, size = 0.1, impr.time = 2)

# sample core based on both genotypes and phenotypes
geno.pheno <- coreHunterData(geno, pheno)
core <- sampleCore(geno.pheno)
}

}
\seealso{
\code{\link{coreHunterData}}, \code{\link{genotypes}},
 \code{\link{phenotypes}}, \code{\link{distances}},
 \code{\link{sampleCore}}, \code{\link{evaluateCore}},
 \code{\link{objective}}
}
\author{
\strong{Maintainer}: Herman De Beukelaer \email{herman.debeukelaer@gmail.com}

Authors:
\itemize{
  \item Guy Davenport \email{daveneti@gmail.com}
}

Other contributors:
\itemize{
  \item Veerle Fack \email{veerle.fack@ugent.be} [thesis advisor]
}

}
\keyword{#'}
\keyword{(Test)}
\keyword{(e.g.}
\keyword{(see}
\keyword{3}
\keyword{8}
\keyword{Core}
\keyword{Hunter}
\keyword{In}
\keyword{Java}
\keyword{Such}
\keyword{The}
\keyword{\url{http://www.corehunter.org}).}
\keyword{a}
\keyword{addition,}
\keyword{allelic}
\keyword{an}
\keyword{and}
\keyword{applications}
\keyword{as}
\keyword{based}
\keyword{be}
\keyword{breeding}
\keyword{bring}
\keyword{can}
\keyword{closer}
\keyword{collections}
\keyword{collections,}
\keyword{construct}
\keyword{core}
\keyword{cores}
\keyword{data,}
\keyword{depending}
\keyword{different}
\keyword{distance}
\keyword{diverse,}
\keyword{diversity,}
\keyword{evaluation}
\keyword{from}
\keyword{general.}
\keyword{genetic}
\keyword{germplasm}
\keyword{have}
\keyword{implemented}
\keyword{in}
\keyword{index}
\keyword{internal}
\keyword{is}
\keyword{large}
\keyword{library}
\keyword{management}
\keyword{many}
\keyword{marker}
\keyword{matrices,}
\keyword{maximum}
\keyword{measures}
\keyword{minimum}
\keyword{multiple}
\keyword{of}
\keyword{on}
\keyword{one}
\keyword{open}
\keyword{optimized}
\keyword{optimizing}
\keyword{or}
\keyword{part}
\keyword{perspectives}
\keyword{phenotypic}
\keyword{plant}
\keyword{precise}
\keyword{precomputed}
\keyword{project}
\keyword{provided}
\keyword{purpose}
\keyword{redundancy.}
\keyword{representative}
\keyword{representativeness,}
\keyword{resource}
\keyword{richness).}
\keyword{sample}
\keyword{simultaneously}
\keyword{so-called}
\keyword{source}
\keyword{subsets}
\keyword{the}
\keyword{to}
\keyword{together.}
\keyword{tool}
\keyword{traits}
\keyword{weighted}
\keyword{with}
