% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{coreHunterData}
\alias{coreHunterData}
\title{Initialize Core Hunter data.}
\usage{
coreHunterData(genotypes, phenotypes, distances)
}
\arguments{
\item{genotypes}{Genetic marker data of class 'chgeno' as returned by function 'genotypes'.}

\item{phenotypes}{Phenotypic trait data of class 'chpheno' as returned by function 'phenotypes'.}

\item{distances}{Precomputed distance matrix of class 'chdist' as returned by function 'distances'.}
}
\description{
The data may contain genotypes, phenotypes and/or a precomputed distance matrix.
}

