% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleData}
\alias{exampleData}
\title{Read example data for 100 individuals.}
\source{
...
}
\usage{
exampleData()
}
\value{
data of class \code{chdata}
}
\description{
Reads a precomputed distance matrix from \code{extdata/distances.csv}.
}

