% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getSize}
\alias{getSize}
\title{Retrieve dataset size.}
\usage{
getSize(data)
}
\arguments{
\item{data}{data object}
}
\value{
dataset size
}
\description{
Retrieve dataset size.
}
\examples{
data <- exampleData()
getSize(data)

}

