% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{objective}
\alias{objective}
\title{Create Core Hunter objective.}
\usage{
objective(type = c("EN", "AN", "EE", "SH", "HE", "CV"), measure = c("PD",
  "GD", "MR", "CE"), weight = 1)
}
\arguments{
\item{type}{...}

\item{measure}{...}

\item{weight}{...}
}
\description{
The following optimization objectives are supported by Core Hunter:
\describe{
 \item{\code{EN}}{
   Average entry-to-nearest-entry distance (default). Maximizes the average distance
   between each selected individual and the closest other selected item
   in the core. Favors diverse cores in which each individual is sufficiently
   different from the most similar other selected item (low redundancy).
   Multiple distance measures are provided to be used with this objective (see below).
 }
 \item{\code{AN}}{
   Average accession-to-nearest-entry distance. Minimizes the average distance
   between each individual (from the full dataset) and the closest selected item
   in the core (which can be the individual itself). Favors representative cores
   in which all items from the original dataset are represented by similar individuals
   in the selected subset. Multiple distance measures are provided to be used with this
   objective (see below).
 }
 \item{\code{EE}}{
   Average entry-to-entry distance. Maximizes the average distance between
   each pair of selected individuals in the core. This objective is related to
   the entry-to-nearest-entry (EN) distance but less effectively avoids redundant,
   similar individuals in the core. In general, use of \code{EN} is preferred.
   Multiple distance measures are provided to be used with this objective (see below).
 }
 \item{\code{SH}}{
   Shannon's allelic diversity index. Maximizes the entropy, as used in information
   theory, of the selected core. Independently takes into account all allele frequencies,
   regardless of the locus (marker) where to which the allele belongs. Requires genotypes.
 }
 \item{\code{HE}}{
   Expected proportion of heterozygous loci. Maximizes the expected proportion of heterzygous
   loci in offspring produced from random crossings within the selected core. In contrast to
   Shannon's index (\code{SH}) this objective treats each marker (locus) with equal importance,
   regardless of the number of possible alleles for that marker. Requires genotypes.
 }
 \item{\code{CV}}{
   Allele coverage. Maximizes the proportion of alleles observed in the full dataset that are
   retained in the selected core. Requires genotoypes.
 }
}
The first three objective types (\code{EN}, \code{AN} and \code{EE}) are based on pairwise distances
between individuals. These distances can be computed using various measures:
\describe{
 \item{\code{PD}}{
   Precomputed distances (default). Use the precomputed distance matrix of the dataset.
 }
 \item{\code{GD}}{
   ...
 }
 \item{\code{MR}}{
   ...
 }
 \item{\code{CE}}{
   ...
 }
}
}

