% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{coreHunterData}
\alias{coreHunterData}
\title{Initialize Core Hunter data.}
\usage{
coreHunterData(distances, genotypes, phenotypes)
}
\arguments{
\item{distances}{Precomputed distance matrix (\code{chdist}).}

\item{genotypes}{Genetic marker data (\code{chgeno}).}

\item{phenotypes}{Phenotypic trait data (\code{chpheno}).}
}
\value{
Core Hunter data (\code{chdata}) with elements \code{dist},
 \code{geno} and/or \code{pheno}, depending on which data is provided,
 as well as an element \code{java} that holds the Java version of the
 data object.
}
\description{
The data may contain genotypes, phenotypes and/or a precomputed distance matrix.
}
\examples{
dist.file <- system.file("extdata", "distances.csv", package = "corehunter")
chData <- coreHunterData(distances = distances(file = dist.file))
chData

}
\seealso{
\code{\link{genotypes}}, \code{\link{phenotypes}}, \code{\link{distances}}
}

