% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{coreHunterData}
\alias{coreHunterData}
\title{Initialize Core Hunter data.}
\usage{
coreHunterData(genotypes, phenotypes, distances)
}
\arguments{
\item{genotypes}{Genetic marker data (\code{chgeno}).}

\item{phenotypes}{Phenotypic trait data (\code{chpheno}).}

\item{distances}{Precomputed distance matrix (\code{chdist}).}
}
\value{
Core Hunter data (\code{chdata}).
}
\description{
The data may contain genotypes, phenotypes and/or a precomputed distance matrix.
}
\seealso{
\code{\link{genotypes}}, \code{\link{phenotypes}}, \code{\link{distances}}
}

