% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{distances}
\alias{distances}
\title{Create distances data from matrix or file.}
\usage{
distances(matrix, file)
}
\arguments{
\item{matrix}{Symmetric distance matrix. Row and column names are required and used as item ids.}

\item{file}{File from which to read the distance matrix.}
}
\value{
distance matrix data (\code{chdist}).
}
\description{
Specify either a symmetric distance matrix or the file from which to read the matrix.
}
\examples{
# read from file

dist.file <- system.file("extdata", "distances.csv", package = "corehunter")
dist <- distances(file = dist.file)

# create from distance matrix

m <- matrix(runif(100), nrow = 10, ncol = 10)
diag(m) <- 0
# make symmetric
m[lower.tri(m)] <- t(m)[lower.tri(m)]
# set headers
rownames(m) <- colnames(m) <- paste("i", 1:10, sep = "-")

dist <- distances(m)
dist

}

