% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluateCore}
\alias{evaluateCore}
\title{Evaluate a core collection using the specified objective.}
\usage{
evaluateCore(core, data, objective)
}
\arguments{
\item{core}{Can be either a core collection reference (\code{chcore}) or a
numeric or character vector indicating the indices or names, respectively,
of the individuals in the evaluated core.}

\item{data}{Core Hunter data (\code{chdata}).}

\item{objective}{Objective function (\code{chobj}) used to evaluate the core.}
}
\value{
Value of the core when evaluated with the chosen objective (numeric).
}
\description{
Evaluate a core collection using the specified objective.
}
\seealso{
\code{\link{coreHunterData}}, \code{\link{objective}}
}

