% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleData}
\alias{exampleData}
\title{Example dataset with 100 individuals.}
\source{
\url{http://bioinformatics.oxfordjournals.org/content/23/16/2155.long}
}
\usage{
exampleData()
}
\value{
Core Hunter data of class \code{chdata}
}
\description{
Reads a precomputed distance matrix from \code{"extdata/distances.csv"},
genotypes from \code{"extdata/genotypes.csv"} and phenotypes from
\code{"extdata/phenotypes.csv"}.
}
\details{
Genotypes and phenotypes are taken from the PowerCore project and, although
this is not really the case, it is assumed here that they both describe
the same population. Only the first 100 individuals are included.
The distance matrix is computed from the phenotypic traits (Gower distance).
}

