% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleData}
\alias{exampleData}
\title{Example dataset with 100 individuals.}
\source{
TODO \url{...} Currently PowerCore but genotypes and phenotypes
  are not really from the same individuals. Looking for better dataset.
}
\usage{
exampleData()
}
\value{
data of class \code{chdata}
}
\description{
Reads a precomputed distance matrix from \code{"extdata/distances.csv"},
genotypes from \code{"extdata/genotypes.csv"} and phenotypes from
\code{"extdata/phenotypes.csv"}.
}

