% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getDistanceMatrix}
\alias{getDistanceMatrix}
\title{Retrieve precomputed distance matrix.}
\usage{
getDistanceMatrix(data)
}
\arguments{
\item{data}{data object containing distances}
}
\value{
distance matrix (numeric)
}
\description{
Retrieve precomputed distance matrix.
}
\examples{
data <- exampleData()
m <- getDistanceMatrix(data)
m

}

