% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{phenotypes}
\alias{phenotypes}
\title{Read phenotype data from file.}
\usage{
phenotypes(file)
}
\arguments{
\item{file}{File containing the phenotype data.}
}
\value{
Phenotype data of class \code{chpheno} with elements
\describe{
 \item{\code{data}}{Phenotypes (data frame).}
 \item{\code{size}}{Number of individuals in the dataset.}
 \item{\code{ids}}{Unique item identifiers.}
 \item{\code{names}}{Item names. Names of individuals to which no explicit name
   has been assigned are equal to the unique \code{ids}.}
 \item{\code{types}}{Variable types and encodings.}
 \item{\code{file}}{Path of file from which data was read (if applicable).}
 \item{\code{java}}{Java version of the data object.}
}
}
\description{
See \url{www.corehunter.org} for documentation and examples
of the phenotype data format used by Core Hunter.
}
\examples{
pheno.file <- system.file("extdata", "phenotypes.csv", package = "corehunter")
pheno <- phenotypes(pheno.file)

}

