% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{sampleCore}
\alias{sampleCore}
\title{Sample a core collection from the given data.}
\usage{
sampleCore(data, size = 0.2, objectives)
}
\arguments{
\item{data}{Core Hunter data of class 'chdata' containing genotypes,
phenotypes and/or a precomputed distance matrix.}

\item{size}{Desired core subset size (numeric). If larger than one the value
is used as the absolute core size after rounding. Else it is used as the sampling
rate and multiplied with the dataset size to determine the size of the core.
The default sampling rate is 0.2.}

\item{objectives}{List of objectives.}
}
\description{
Sample a core collection from the given data.
}

